+++ .cspell.json
@@ -34,6 +34,13 @@
         "BOOTLOADER",
         "devcontainer",
         "gasleft",
-        "chainid"
+        "chainid",
+        "myfilebase",
+        "ethersproject",
+        "Numberish",
+        "Blockhash",
+        "Typechain",
+        "Soulbound",
+        "accesscontrolrole"
     ]
 }
