+++ subquery/testnet-complete.ts
@@ -36,63 +36,14 @@ const project: EthereumProject = {
       kind: EthereumDatasourceKind.Runtime,
       startBlock: 1993364, // This is the block that the contract was deployed on
       options: {
-        abi: "erc721",
+        abi: "ClickContentSign",
         address: "0x999368030Ba79898E83EaAE0E49E89B7f6410940",
       },
       assets: new Map([
         [
-          "erc721",
+          "ClickContentSign",
           {
-            file: "./abis/erc721.abi.json",
-          },
-        ],
-      ]),
-      mapping: {
-        file: "./dist/index.js",
-        handlers: [
-          {
-            kind: EthereumHandlerKind.Call,
-            handler: "handleSafeMint",
-            filter: {
-              function: "safeMint(address,string)",
-            },
-          },
-          {
-            kind: EthereumHandlerKind.Event,
-            handler: "handleApproval",
-            filter: {
-              topics: ["Approval(address,address,uint256)"],
-            },
-          },
-          {
-            kind: EthereumHandlerKind.Event,
-            handler: "handleApprovalForAll",
-            filter: {
-              topics: ["ApprovalForAll(address,address,bool)"],
-            },
-          },
-          {
-            kind: EthereumHandlerKind.Event,
-            handler: "handleTransfer",
-            filter: {
-              topics: ["Transfer(address,address,uint256)"],
-            },
-          },
-        ],
-      },
-    },
-    {
-      kind: EthereumDatasourceKind.Runtime,
-      startBlock: 1993364, // This is the block that the contract was deployed on
-      options: {
-        abi: "erc721",
-        address: "0x195e4E251c41e8Ae9E9E961366C73e2CFbfB115A",
-      },
-      assets: new Map([
-        [
-          "erc721",
-          {
-            file: "./abis/erc721.abi.json",
+            file: "./abis/ClickContentSign.abi.json",
           },
         ],
       ]),
@@ -134,14 +85,14 @@ const project: EthereumProject = {
       kind: EthereumDatasourceKind.Runtime,
       startBlock: 2178049, // This is the block that the contract was deployed on
       options: {
-        abi: "erc20",
+        abi: "NODL",
         address: "0xb4B74C2BfeA877672B938E408Bae8894918fE41C",
       },
       assets: new Map([
         [
-          "erc20",
+          "NODL",
           {
-            file: "./abis/erc20.abi.json",
+            file: "./abis/NODL.abi.json",
           },
         ],
       ]),
@@ -169,14 +120,14 @@ const project: EthereumProject = {
       kind: EthereumDatasourceKind.Runtime,
       startBlock: 2157320,
       options: {
-        abi: "migration",
+        abi: "NODLMigration",
         address: "0x1427d38B967435a3F8f476Cda0bc4F51fe66AF4D",
       },
       assets: new Map([
         [
-          "migration",
+          "NODLMigration",
           {
-            file: "./abis/migration.abi.json",
+            file: "./abis/NODLMigration.abi.json",
           },
         ],
       ]),
@@ -211,14 +162,14 @@ const project: EthereumProject = {
       kind: EthereumDatasourceKind.Runtime,
       startBlock: 3001690, // This is the block that the contract was deployed on
       options: {
-        abi: "erc721-a",
+        abi: "MigrationNFT",
         address: "0x9Fed2d216DBE36928613812400Fd1B812f118438",
       },
       assets: new Map([
         [
-          "erc721-a",
+          "MigrationNFT",
           {
-            file: "./abis/erc721-a.abi.json",
+            file: "./abis/MigrationNFT.abi.json",
           },
         ],
       ]),
@@ -240,14 +191,14 @@ const project: EthereumProject = {
       kind: EthereumDatasourceKind.Runtime,
       startBlock: 3548139, // This is the block that the contract was deployed on
       options: {
-        abi: "grants",
+        abi: "Grants",
         address: "0xdAdF329E8b30D878b139074De163D3A591aAB394",
       },
       assets: new Map([
         [
-          "grants",
+          "Grants",
           {
-            file: "./abis/grants.abi.json",
+            file: "./abis/Grants.abi.json",
           },
         ],
       ]),
@@ -258,14 +209,14 @@ const project: EthereumProject = {
             kind: EthereumHandlerKind.Event,
             handler: "handleClaimed",
             filter: {
-              topics: ["Claimed(address, uint256)"],
+              topics: ["Claimed(address, uint256, uint256, uint256)"],
             },
           },
           {
             kind: EthereumHandlerKind.Event,
             handler: "handleRenounced",
             filter: {
-              topics: ["Renounced(address, address)"],
+              topics: ["Renounced(address, address, uint256, uint256)"],
             },
           },
           {
@@ -281,7 +232,7 @@ const project: EthereumProject = {
             kind: EthereumHandlerKind.Event,
             handler: "handleVestingSchedulesCanceled",
             filter: {
-              topics: ["VestingSchedulesCanceled(address, address)"],
+              topics: ["VestingSchedulesCanceled(address, address, uint256, uint256)"],
             },
           },
         ],
@@ -291,14 +242,14 @@ const project: EthereumProject = {
       kind: EthereumDatasourceKind.Runtime,
       startBlock: 3548120, // This is the block that the contract was deployed on
       options: {
-        abi: "grantsMigration",
+        abi: "GrantsMigration",
         address: "0xF81b3b954221BeDcf762cd18FEc1A22D25016B2E",
       },
       assets: new Map([
         [
-          "grantsMigration",
+          "GrantsMigration",
           {
-            file: "./abis/grantsMigration.abi.json",
+            file: "./abis/GrantsMigration.abi.json",
           },
         ],
       ]),
@@ -333,14 +284,14 @@ const project: EthereumProject = {
       kind: EthereumDatasourceKind.Runtime,
       startBlock: 3798170, // This is the block that the contract was deployed on
       options: {
-        abi: "rewards",
+        abi: "Rewards",
         address: "0xba8a8ff7E7332f7e05205Ec9fC927965435C552c",
       },
       assets: new Map([
         [
-          "rewards",
+          "Rewards",
           {
-            file: "./abis/rewards.abi.json",
+            file: "./abis/Rewards.abi.json",
           },
         ],
       ]),
